Move the new onboarding stuff inside the to be replaced IdeaSuggestionBox
