+++ webview-ui/src/components/welcome/WelcomeViewProvider.tsx
@@ -22,11 +22,6 @@ import RooHero from "./RooHero"
 import { Trans } from "react-i18next"
 import { ArrowLeft, ArrowRight, BadgeInfo, Brain, TriangleAlert } from "lucide-react"
 import { buildDocLink } from "@/utils/docLinks"
-// kilocode_change start: Onboarding screens
-import { NoFolderNoHistoryScreen } from "./screens/NoFolderNoHistoryScreen"
-import { FolderNoHistoryScreen } from "./screens/FolderNoHistoryScreen"
-import { PaidModelScreen } from "./screens/PaidModelScreen"
-// kilocode_change end
 
 type ProviderOption = "roo" | "custom"
 type AuthOrigin = "landing" | "providerSelection"
@@ -49,34 +44,6 @@ const WelcomeViewProvider = () => {
 	const [manualUrl, setManualUrl] = useState("")
 	const [manualErrorMessage, setManualErrorMessage] = useState<boolean | undefined>(undefined)
 	const manualUrlInputRef = useRef<HTMLInputElement | null>(null)
-	// kilocode_change start: Onboarding state
-	const [hasOpenFolder, setHasOpenFolder] = useState<boolean>(false)
-	const [hasSessionHistory, setHasSessionHistory] = useState<boolean>(false)
-	const [showPaidModelScreen, _setShowPaidModelScreen] = useState<boolean>(false)
-	const [supportsKiloGateway, _setSupportsKiloGateway] = useState<boolean>(false)
-
-	// Check workspace state on mount
-	useEffect(() => {
-		vscode.postMessage({ type: "checkWorkspaceState" })
-	}, [])
-
-	// Listen for workspace state updates
-	useEffect(() => {
-		const handler = (event: MessageEvent) => {
-			const message = event.data
-
-			switch (message.type) {
-				case "workspaceState":
-					setHasOpenFolder(message.hasFolder)
-					setHasSessionHistory(message.hasHistory)
-					break
-			}
-		}
-
-		window.addEventListener("message", handler)
-		return () => window.removeEventListener("message", handler)
-	}, [])
-	// kilocode_change end
 
 	// When auth completes during the provider signup flow, either:
 	// 1. If user skipped model selection (cloudAuthSkipModel=true), navigate to provider selection with "custom" selected
@@ -215,23 +182,6 @@ const WelcomeViewProvider = () => {
 	const handleOpenSignupUrl = () => {
 		vscode.postMessage({ type: "rooCloudSignIn", useProviderSignup: false })
 	}
-
-	// kilocode_change start: Onboarding flow routing
-	// Initial load - no folder, no history
-	if (!hasOpenFolder && !hasSessionHistory && selectedProvider === null) {
-		return <NoFolderNoHistoryScreen />
-	}
-
-	// Has folder but no history
-	if (hasOpenFolder && !hasSessionHistory && selectedProvider === null) {
-		return <FolderNoHistoryScreen />
-	}
-
-	// Show paid model screen if triggered
-	if (showPaidModelScreen) {
-		return <PaidModelScreen supportsKiloGateway={supportsKiloGateway} />
-	}
-	// kilocode_change end
 
 	// Render the waiting for cloud state
 	if (authInProgress) {
